refactor: move urlsForCards into lib
