+++ src/behavior.test.js
@@ -1,4 +1,4 @@
-const { urlsForCards, replyWithCards } = require('./behavior')
+const { replyWithCards } = require('./behavior')
 
 const cardDb = {
   'fire sigil': {
@@ -12,50 +12,8 @@ const cardDb = {
   }
 }
 
-describe('urlsForCards', () => {
-  describe('single fetch', () => {
-    const cards = ['fire sigil']
-    it('returns one url', () => {
-      expect(urlsForCards(cards, cardDb)).toEqual({
-        urls: ['firesigil.html'],
-        errors: []
-      })
-    })
-  })
-
-  describe('triple fetch', () => {
-    const cards = ['fire sigil', 'time sigil', 'wisdom of the elders']
-    it('returns all 3 urls', () => {
-      expect(urlsForCards(cards, cardDb)).toEqual({
-        urls: ['firesigil.html', 'timesigil.html', 'wisdom.html'],
-        errors: []
-      })
-    })
-  })
-
-  describe('bad fetch', () => {
-    const cards = ['not a card']
-    it('fills error array', () => {
-      expect(urlsForCards(cards, cardDb)).toEqual({
-        urls: [],
-        errors: ['not a card']
-      })
-    })
-  })
-
-  describe('bad and good fetch', () => {
-    const cards = ['time sigil', 'not a card', 'fire sigil']
-    it('only good goes through', () => {
-      expect(urlsForCards(cards, cardDb)).toEqual({
-        urls: ['timesigil.html', 'firesigil.html'],
-        errors: ['not a card']
-      })
-    })
-  })
-})
-
 describe('replyWithCards', () => {
-  describe('does not answer none', () => {
+  it('does not answer none', () => {
     expect(
       replyWithCards({
         msgText: 'One With Nothing is the best MTG card',
@@ -64,7 +22,7 @@ describe('replyWithCards', () => {
     ).toEqual([])
   })
 
-  describe('answers one', () => {
+  it('answers one', () => {
     expect(
       replyWithCards({
         msgText: 'Start with {{Fire Sigil}}',
@@ -73,7 +31,7 @@ describe('replyWithCards', () => {
     ).toEqual(['firesigil.html'])
   })
 
-  describe('answers two', () => {
+  it('answers two', () => {
     expect(
       replyWithCards({
         msgText: 'Then play {{Fire Sigil}} and {{Wisdom of the Elders}}',
@@ -82,7 +40,7 @@ describe('replyWithCards', () => {
     ).toEqual(['firesigil.html wisdom.html'])
   })
 
-  describe('answers errors', () => {
+  it('answers errors', () => {
     expect(
       replyWithCards({
         msgText: 'Next play {{Time Walk}} and {{Black Lotus}}',
@@ -91,7 +49,7 @@ describe('replyWithCards', () => {
     ).toEqual(['Could not find any cards named time walk, black lotus'])
   })
 
-  describe('answers everything', () => {
+  it('answers everything', () => {
     expect(
       replyWithCards({
         msgText:
