+++ src/lib.test.js
@@ -2,7 +2,8 @@ const {
   cardsByName,
   requestedCards,
   fetchUniqueName,
-  sanitize
+  sanitize,
+  urlsForCards
 } = require('./lib')
 
 describe('cardsByName', () => {
@@ -108,5 +109,59 @@ describe('sanitize', () => {
     expect(sanitize('   WISDOM Of the* \t Elders ')).toEqual(
       'wisdom of the elders'
     )
+  })
+})
+
+describe('urlsForCards', () => {
+  const cardDb = {
+    'fire sigil': {
+      DetailsUrl: 'firesigil.html'
+    },
+    'time sigil': {
+      DetailsUrl: 'timesigil.html'
+    },
+    'wisdom of the elders': {
+      DetailsUrl: 'wisdom.html'
+    }
+  }
+
+  describe('single fetch', () => {
+    const cards = ['fire sigil']
+    it('returns one url', () => {
+      expect(urlsForCards(cards, cardDb)).toEqual({
+        urls: ['firesigil.html'],
+        errors: []
+      })
+    })
+  })
+
+  describe('triple fetch', () => {
+    const cards = ['fire sigil', 'time sigil', 'wisdom of the elders']
+    it('returns all 3 urls', () => {
+      expect(urlsForCards(cards, cardDb)).toEqual({
+        urls: ['firesigil.html', 'timesigil.html', 'wisdom.html'],
+        errors: []
+      })
+    })
+  })
+
+  describe('bad fetch', () => {
+    const cards = ['not a card']
+    it('fills error array', () => {
+      expect(urlsForCards(cards, cardDb)).toEqual({
+        urls: [],
+        errors: ['not a card']
+      })
+    })
+  })
+
+  describe('bad and good fetch', () => {
+    const cards = ['time sigil', 'not a card', 'fire sigil']
+    it('only good goes through', () => {
+      expect(urlsForCards(cards, cardDb)).toEqual({
+        urls: ['timesigil.html', 'firesigil.html'],
+        errors: ['not a card']
+      })
+    })
   })
 })
