+++ src/main.js
@@ -1,8 +1,9 @@
 require('dotenv').config()
 
 const { cardsByName } = require('./lib')
-const { onMessageCreate } = require('behavior')
-const Eris = require('eris')
+const { replyWithCards } = require('./behavior')
+const { createBot } = require('./functional_bot')
+
 const fs = require('fs')
 
 const CARD_DATA_SOURCE = 'assets/eternal-cards-1.38.json'
@@ -11,12 +12,9 @@ const cardData = JSON.parse(fs.readFileSync(CARD_DATA_SOURCE))
 const cardDb = cardsByName(cardData)
 console.log(`${Object.keys(cardDb).length} cards loaded`)
 
-const bot = new Eris(process.env.DISCORD_BOT_TOKEN)
-
-bot.on('ready', () => console.log('Ready!'))
-
-bot.on('messageCreate', incomingMsg => {
-  onMessageCreate(bot, incomingMsg, cardDb)
+const bot = createBot({
+  discordBotToken: process.env.DISCORD_BOT_TOKEN,
+  onReady: () => console.log('Aegiseek is connected!'),
+  responsesForMessage: msg => replyWithCards({ msgText: msg.content, cardDb })
 })
-
 bot.connect()
